fix: centralize checkpoint integrity guards in core
